add a github ci workflow:
test: poetry run pytest -v
test-builds: docker buildx build . --build-arg PYTHON_VERSION=3.12.3

The latter can be run against any number of directories containing
test results.  As these tests are not as quick as pytest, and
hopefully in the near future there will be dozens if not more such
directories; it makes sense to define a criteria for which
directories to test.  Perhaps something along the line of:
  * Dockerfile is somewhat unique from a docker perspective (example:
    makes use of caching)
  * Specific test has been proven to be problematic or regression prone.

Over time, more tests can be added.  For example Rails has:
  * System test: https://github.com/fly-apps/dockerfile-rails/blob/main/.github/workflows/system-test.yml
  * Linting: https://github.com/fly-apps/dockerfile-rails/blob/main/.github/workflows/rubocop.yml