Unpin 'bullseye' as the Debian version.
For starters, there is no python version 3.12.3-slim-bookworm; there is a python
version 3.12.3, and there is a docker image with the tag 3.12.3-slim-bookworm

The complete set of Python images can be found here: https://hub.docker.com/_/python

Images are based on specific Linux versions: debian full, debian slim, and alpine.
For Debian, the default is the current stable release at the time the image was
created, or selected named releases.  The list of names, and dates, can be found
here:

https://wiki.debian.org/DebianReleases#Production_Releases

At the present time, bookworm is the stable release, and bullseye is the previous
stable release.  New Python images are make with both, but presumably that will
stop sometime around the end of life for bullseye, which is scheduled for July,
at which point `fly launch` with newer releases of Python made after this point
will fail.

The solution is to star with the version of Debian selected by those that package
up Python images for each release of Python.  Should there be a need, additional
dockerfile generator options can be created to specify which image is desired.
dockerfile-rails supports --alpine and --variant options; it also supports
--fullstaq and --jemalloc which are hosted on places other than dockerhub.

Such options can be handled later, if and when the need applies.